refactoring sidenav service
