+++ src/app/components/sidenav/sidenav.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { RoutesService } from 'src/app/services/routes/routes.service';
-import { SidenavToggleService } from 'src/app/sidenav-toggle.service';
+import { SidenavToggleService } from 'src/app/services/sidenav-toggle/sidenav-toggle.service';
 
 @Component({
   selector: 'app-sidenav',
