+++ src/app/services/tfr-management/tfr-management.service.ts
@@ -1,16 +1,15 @@
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, catchError, of, throwError } from 'rxjs';
 import {
   getAllocatedResourcesURL,
   getProjectURL,
   getUpdateProjectStatusURL,
 } from 'src/app/shared/utils';
-import { Milestone, Project } from 'src/app/shared/interfaces';
-
-import { catchError, of, throwError } from 'rxjs';
 import { resourceProjectsURL, projectsURL } from 'src/app/shared/constants';
 import {
+  Milestone,
+  Project,
   AllocatedResourceTypeDTO,
   ProjectBasicDetails,
   ProjectResourceDTO,
